Support aliases for channels
It is possible to declare a 'Channels' section in the configuration
file (telepaatti.conf in /etc or in the current directory) for defining
channel aliases. This has the form:

    [Channels]
    jdev: jdev@conference.jabberfr.org
    foo: welovefoo@chat.example.org
    (...)

The users can then join a MUC by joining #jdev or #foo without having
to give the whole JID of the room.